revised tests and revised functions because the test were not passing at first
